Extract ProcessListTable from ProcessList panel
